+++ spacepackets/cfdp/__init__.py
@@ -22,6 +22,8 @@ from .tlv import (
     FaultHandlerOverrideTlv,
     FilestoreActionCode,
     FilestoreResponseStatusCode,
+    TlvHolder,
+    TlvTypeMissmatch,
 )
 from .lv import CfdpLv
 from .conf import PduConfig
